Updated naming scheme for load & smoke tests, updated jenkinsfile to delete old test deployment, updated deployments to allow 1 max unavailable.
