+++ smoke-test-blue.js
@@ -2,15 +2,16 @@ import { sleep } from 'k6';
 import http from 'k6/http';
 
 export const options = {
-    duration: '2m',
-    vus: 20,
+    iterations: 5,
+    vus: 5,
 };
 
 export default function () {
+    let body = {email: 'testing@test.test', password: 'TestTest1!'}
 
-
-    http.get('http://a8164654749db438791fab617e96f1c9-157527310.us-east-1.elb.amazonaws.com:5000/api/product');
-    sleep(0.01);
-    http.get('http://a8164654749db438791fab617e96f1c9-157527310.us-east-1.elb.amazonaws.com:5000/api/product');
-    sleep(0.01);
+    http.post('http://a8164654749db438791fab617e96f1c9-157527310.us-east-1.elb.amazonaws.com:5000/auth/login',
+    JSON.stringify(body));
+    sleep(0.1)
+    http.post('http://a8164654749db438791fab617e96f1c9-157527310.us-east-1.elb.amazonaws.com:5000/auth/logout',
+    JSON.stringify(''))
 }
